commit6

+++ Core/Inc/TanControl.h
@@ -3,20 +3,20 @@
 #include <cmath>
 
 struct Spd{
-	int32_t lin;
-	int32_t ang;
+	float lin;
+	float ang;
 };
 
 class TanControl{
 public:
-	Spd calcTgtRobotSpds(int32_t dist, int32_t ang) {
+	Spd calcTgtRobotSpds(float dist, float ang) {
 		clearSpd();
 		spd.lin = MAX_MOT_SPEED * tanh(dist) * cos(ang);
 		spd.ang = K_ANG * ang + spd.lin * sin(ang) / dist;
 		return spd;
 	}
 
-	int32_t getErrorAngle(float *distances){
+	float getErrorAngle(float *distances){
 		int32_t errorAngle = 0;
 		int32_t minDistance = 10000;
 		  for(int i = 0;i<=360;i++){
@@ -29,7 +29,7 @@ public:
 		  return errorAngle;
 	}
 private:
-	int32_t limitAng(int32_t ang){
+	float limitAng(float ang){
 		int32_t inAng = ang;
 		while (inAng >= HALF_CIRCLE) {
 			inAng -= 2 * HALF_CIRCLE;
